Added tophat module
